# Refactor probing logic into an external iterator