@cbarrick 2020-07-26
Also, I noticed that all of the iterator methods are marked `unsafe` with comments like:

> ```
> /// It is up to the caller to ensure that the `RawTable` outlives the `RawDrain`.
> /// Because we cannot make the `next` method unsafe on the `RawDrain`,
> /// we have to make the `drain` method unsafe.
> ```

I followed the same pattern, but I have some doubts.

But the `RawDrain` and the new `RawIterHash` both contain lifetime parameters that tie them to the table, and the `RawIntoIter` owns the table. So I think the methods that create these iterators are actually safe.