@Amanieu 2020-07-26
Did you run the benchmarks to check that there are no regressions to the performance of `find`? This is the single most performance-critical function of `hashbrown`.