@cbarrick 2020-07-26
I believe it's the exact same logic, except that the pattern matching in the outer for-loop has been replaced by `pos.unwrap()`. So I don't expect a regression.

I will run the benchmarks on my 2013 MBP. I may be able to test on a Ryzen 2500X too.