@cbarrick 2020-07-26
# 2013 MacBook Pro

It doesn't look like there's any change. The results were often better but within the margin of error.

- Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
- MacOS Catalina 10.15.6

## Side-by-Side

Benchmark                        | Before                      | After
---------------------------------|-----------------------------|-----------------------------
test clone_from_large            |  2,455 ns/iter (+/- 91)     |  2,448 ns/iter (+/- 77)
test clone_from_small            |     35 ns/iter (+/- 1)      |     35 ns/iter (+/- 1)
test clone_large                 |  2,863 ns/iter (+/- 162)    |  2,986 ns/iter (+/- 584)
test clone_small                 |    122 ns/iter (+/- 3)      |    120 ns/iter (+/- 11)
test insert_ahash_highbits       | 12,132 ns/iter (+/- 3,023)  | 10,656 ns/iter (+/- 658)
test insert_ahash_random         | 12,180 ns/iter (+/- 280)    | 11,409 ns/iter (+/- 319)
test insert_ahash_serial         | 10,674 ns/iter (+/- 223)    | 10,577 ns/iter (+/- 263)
test insert_erase_ahash_highbits | 25,289 ns/iter (+/- 1,277)  | 25,772 ns/iter (+/- 11,353)
test insert_erase_ahash_random   | 26,840 ns/iter (+/- 7,069)  | 25,946 ns/iter (+/- 710)
test insert_erase_ahash_serial   | 24,171 ns/iter (+/- 1,263)  | 26,268 ns/iter (+/- 7,075)
test insert_erase_std_highbits   | 61,404 ns/iter (+/- 19,648) | 60,708 ns/iter (+/- 1,436)
test insert_erase_std_random     | 60,076 ns/iter (+/- 10,963) | 62,395 ns/iter (+/- 21,598)
test insert_erase_std_serial     | 52,745 ns/iter (+/- 15,662) | 54,522 ns/iter (+/- 2,000)
test insert_std_highbits         | 27,610 ns/iter (+/- 695)    | 27,265 ns/iter (+/- 756)
test insert_std_random           | 28,313 ns/iter (+/- 1,170)  | 27,059 ns/iter (+/- 409)
test insert_std_serial           | 24,097 ns/iter (+/- 579)    | 24,720 ns/iter (+/- 737)
test iter_ahash_highbits         |  1,801 ns/iter (+/- 295)    |  2,153 ns/iter (+/- 58)
test iter_ahash_random           |  1,715 ns/iter (+/- 69)     |  1,929 ns/iter (+/- 143)
test iter_ahash_serial           |  2,021 ns/iter (+/- 55)     |  1,965 ns/iter (+/- 118)
test iter_std_highbits           |  1,756 ns/iter (+/- 129)    |  2,018 ns/iter (+/- 65)
test iter_std_random             |  1,720 ns/iter (+/- 46)     |  1,897 ns/iter (+/- 170)
test iter_std_serial             |  1,920 ns/iter (+/- 284)    |  1,770 ns/iter (+/- 87)
test lookup_ahash_highbits       |  5,227 ns/iter (+/- 1,501)  |  4,419 ns/iter (+/- 871)
test lookup_ahash_random         |  5,320 ns/iter (+/- 76)     |  4,313 ns/iter (+/- 193)
test lookup_ahash_serial         |  4,592 ns/iter (+/- 88)     |  3,968 ns/iter (+/- 88)
test lookup_fail_ahash_highbits  |  4,337 ns/iter (+/- 97)     |  3,851 ns/iter (+/- 97)
test lookup_fail_ahash_random    |  5,044 ns/iter (+/- 103)    |  4,700 ns/iter (+/- 1,866)
test lookup_fail_ahash_serial    |  3,885 ns/iter (+/- 132)    |  3,440 ns/iter (+/- 375)
test lookup_fail_std_highbits    | 21,981 ns/iter (+/- 1,649)  | 21,550 ns/iter (+/- 536)
test lookup_fail_std_random      | 22,589 ns/iter (+/- 5,586)  | 21,631 ns/iter (+/- 473)
test lookup_fail_std_serial      | 18,945 ns/iter (+/- 419)    | 18,595 ns/iter (+/- 6,583)
test lookup_std_highbits         | 21,980 ns/iter (+/- 509)    | 21,396 ns/iter (+/- 619)
test lookup_std_random           | 22,355 ns/iter (+/- 403)    | 21,246 ns/iter (+/- 16,381)
test lookup_std_serial           | 19,415 ns/iter (+/- 346)    | 18,689 ns/iter (+/- 566)


## Before

```
test clone_from_large            ... bench:       2,455 ns/iter (+/- 91)     
test clone_from_small            ... bench:          35 ns/iter (+/- 1)      
test clone_large                 ... bench:       2,863 ns/iter (+/- 162)    
test clone_small                 ... bench:         122 ns/iter (+/- 3)      
test insert_ahash_highbits       ... bench:      12,132 ns/iter (+/- 3,023)  
test insert_ahash_random         ... bench:      12,180 ns/iter (+/- 280)    
test insert_ahash_serial         ... bench:      10,674 ns/iter (+/- 223)    
test insert_erase_ahash_highbits ... bench:      25,289 ns/iter (+/- 1,277)  
test insert_erase_ahash_random   ... bench:      26,840 ns/iter (+/- 7,069)  
test insert_erase_ahash_serial   ... bench:      24,171 ns/iter (+/- 1,263)  
test insert_erase_std_highbits   ... bench:      61,404 ns/iter (+/- 19,648) 
test insert_erase_std_random     ... bench:      60,076 ns/iter (+/- 10,963) 
test insert_erase_std_serial     ... bench:      52,745 ns/iter (+/- 15,662) 
test insert_std_highbits         ... bench:      27,610 ns/iter (+/- 695)    
test insert_std_random           ... bench:      28,313 ns/iter (+/- 1,170)  
test insert_std_serial           ... bench:      24,097 ns/iter (+/- 579)    
test iter_ahash_highbits         ... bench:       1,801 ns/iter (+/- 295)    
test iter_ahash_random           ... bench:       1,715 ns/iter (+/- 69)     
test iter_ahash_serial           ... bench:       2,021 ns/iter (+/- 55)     
test iter_std_highbits           ... bench:       1,756 ns/iter (+/- 129)    
test iter_std_random             ... bench:       1,720 ns/iter (+/- 46)     
test iter_std_serial             ... bench:       1,920 ns/iter (+/- 284)    
test lookup_ahash_highbits       ... bench:       5,227 ns/iter (+/- 1,501)  
test lookup_ahash_random         ... bench:       5,320 ns/iter (+/- 76)     
test lookup_ahash_serial         ... bench:       4,592 ns/iter (+/- 88)     
test lookup_fail_ahash_highbits  ... bench:       4,337 ns/iter (+/- 97)     
test lookup_fail_ahash_random    ... bench:       5,044 ns/iter (+/- 103)    
test lookup_fail_ahash_serial    ... bench:       3,885 ns/iter (+/- 132)    
test lookup_fail_std_highbits    ... bench:      21,981 ns/iter (+/- 1,649)  
test lookup_fail_std_random      ... bench:      22,589 ns/iter (+/- 5,586)  
test lookup_fail_std_serial      ... bench:      18,945 ns/iter (+/- 419)    
test lookup_std_highbits         ... bench:      21,980 ns/iter (+/- 509)    
test lookup_std_random           ... bench:      22,355 ns/iter (+/- 403)    
test lookup_std_serial           ... bench:      19,415 ns/iter (+/- 346)    
```


## After

```
test clone_from_large            ... bench:       2,448 ns/iter (+/- 77)
test clone_from_small            ... bench:          35 ns/iter (+/- 1)
test clone_large                 ... bench:       2,986 ns/iter (+/- 584)
test clone_small                 ... bench:         120 ns/iter (+/- 11)
test insert_ahash_highbits       ... bench:      10,656 ns/iter (+/- 658)
test insert_ahash_random         ... bench:      11,409 ns/iter (+/- 319)
test insert_ahash_serial         ... bench:      10,577 ns/iter (+/- 263)
test insert_erase_ahash_highbits ... bench:      25,772 ns/iter (+/- 11,353)
test insert_erase_ahash_random   ... bench:      25,946 ns/iter (+/- 710)
test insert_erase_ahash_serial   ... bench:      26,268 ns/iter (+/- 7,075)
test insert_erase_std_highbits   ... bench:      60,708 ns/iter (+/- 1,436)
test insert_erase_std_random     ... bench:      62,395 ns/iter (+/- 21,598)
test insert_erase_std_serial     ... bench:      54,522 ns/iter (+/- 2,000)
test insert_std_highbits         ... bench:      27,265 ns/iter (+/- 756)
test insert_std_random           ... bench:      27,059 ns/iter (+/- 409)
test insert_std_serial           ... bench:      24,720 ns/iter (+/- 737)
test iter_ahash_highbits         ... bench:       2,153 ns/iter (+/- 58)
test iter_ahash_random           ... bench:       1,929 ns/iter (+/- 143)
test iter_ahash_serial           ... bench:       1,965 ns/iter (+/- 118)
test iter_std_highbits           ... bench:       2,018 ns/iter (+/- 65)
test iter_std_random             ... bench:       1,897 ns/iter (+/- 170)
test iter_std_serial             ... bench:       1,770 ns/iter (+/- 87)
test lookup_ahash_highbits       ... bench:       4,419 ns/iter (+/- 871)
test lookup_ahash_random         ... bench:       4,313 ns/iter (+/- 193)
test lookup_ahash_serial         ... bench:       3,968 ns/iter (+/- 88)
test lookup_fail_ahash_highbits  ... bench:       3,851 ns/iter (+/- 97)
test lookup_fail_ahash_random    ... bench:       4,700 ns/iter (+/- 1,866)
test lookup_fail_ahash_serial    ... bench:       3,440 ns/iter (+/- 375)
test lookup_fail_std_highbits    ... bench:      21,550 ns/iter (+/- 536)
test lookup_fail_std_random      ... bench:      21,631 ns/iter (+/- 473)
test lookup_fail_std_serial      ... bench:      18,595 ns/iter (+/- 6,583)
test lookup_std_highbits         ... bench:      21,396 ns/iter (+/- 619)
test lookup_std_random           ... bench:      21,246 ns/iter (+/- 16,381)
test lookup_std_serial           ... bench:      18,689 ns/iter (+/- 566)
```