@cbarrick 2020-07-26
# Linux on Ryzen 1600X

Again no change. If anything, lookup performance may be slightly better.

- AMD Ryzen 5 1600X
- Debian Sid (2020-07-26)
- rustc 1.47.0-nightly (d6953df14 2020-07-25)


## Side-by-Side

 Benchmark                       | Before                      | After
---------------------------------|-----------------------------|---------------------------
test clone_from_large            |  2,039 ns/iter (+/- 121)    |  2,170 ns/iter (+/- 11)
test clone_from_small            |     22 ns/iter (+/- 0)      |     22 ns/iter (+/- 0)
test clone_large                 |  2,083 ns/iter (+/- 56)     |  2,137 ns/iter (+/- 60)
test clone_small                 |     33 ns/iter (+/- 0)      |     34 ns/iter (+/- 0)
test insert_ahash_highbits       |  6,415 ns/iter (+/- 11)     |  6,179 ns/iter (+/- 6)
test insert_ahash_random         |  7,067 ns/iter (+/- 18)     |  6,888 ns/iter (+/- 53)
test insert_ahash_serial         |  6,472 ns/iter (+/- 36)     |  6,318 ns/iter (+/- 18)
test insert_erase_ahash_highbits | 16,796 ns/iter (+/- 30)     | 16,504 ns/iter (+/- 29)
test insert_erase_ahash_random   | 19,554 ns/iter (+/- 243)    | 19,412 ns/iter (+/- 33)
test insert_erase_ahash_serial   | 16,612 ns/iter (+/- 34)     | 15,993 ns/iter (+/- 41)
test insert_erase_std_highbits   | 42,649 ns/iter (+/- 101)    | 44,060 ns/iter (+/- 108)
test insert_erase_std_random     | 44,372 ns/iter (+/- 10,354) | 44,821 ns/iter (+/- 1,881)
test insert_erase_std_serial     | 37,117 ns/iter (+/- 275)    | 36,760 ns/iter (+/- 34)
test insert_std_highbits         | 18,120 ns/iter (+/- 1,112)  | 18,071 ns/iter (+/- 933)
test insert_std_random           | 18,487 ns/iter (+/- 735)    | 18,021 ns/iter (+/- 300)
test insert_std_serial           | 15,781 ns/iter (+/- 368)    | 16,315 ns/iter (+/- 101)
test iter_ahash_highbits         |  1,830 ns/iter (+/- 90)     |  1,854 ns/iter (+/- 37)
test iter_ahash_random           |  1,966 ns/iter (+/- 28)     |  1,911 ns/iter (+/- 2)
test iter_ahash_serial           |  1,768 ns/iter (+/- 8)      |  1,795 ns/iter (+/- 72)
test iter_std_highbits           |  1,928 ns/iter (+/- 48)     |  1,901 ns/iter (+/- 33)
test iter_std_random             |  1,961 ns/iter (+/- 113)    |  1,927 ns/iter (+/- 15)
test iter_std_serial             |  1,856 ns/iter (+/- 10)     |  1,842 ns/iter (+/- 64)
test lookup_ahash_highbits       |  3,918 ns/iter (+/- 84)     |  3,320 ns/iter (+/- 4)
test lookup_ahash_random         |  3,974 ns/iter (+/- 8)      |  3,513 ns/iter (+/- 4)
test lookup_ahash_serial         |  3,879 ns/iter (+/- 180)    |  3,113 ns/iter (+/- 4)
test lookup_fail_ahash_highbits  |  2,714 ns/iter (+/- 10)     |  2,669 ns/iter (+/- 13)
test lookup_fail_ahash_random    |  3,644 ns/iter (+/- 70)     |  3,633 ns/iter (+/- 181)
test lookup_fail_ahash_serial    |  2,739 ns/iter (+/- 31)     |  2,653 ns/iter (+/- 24)
test lookup_fail_std_highbits    | 15,765 ns/iter (+/- 711)    | 15,425 ns/iter (+/- 53)
test lookup_fail_std_random      | 15,766 ns/iter (+/- 63)     | 15,293 ns/iter (+/- 85)
test lookup_fail_std_serial      | 14,495 ns/iter (+/- 93)     | 13,399 ns/iter (+/- 83)
test lookup_std_highbits         | 15,337 ns/iter (+/- 707)    | 15,322 ns/iter (+/- 303)
test lookup_std_random           | 15,370 ns/iter (+/- 408)    | 15,372 ns/iter (+/- 97)
test lookup_std_serial           | 14,087 ns/iter (+/- 1,303)  | 13,525 ns/iter (+/- 557)

## Before

```
test clone_from_large            ... bench:       2,039 ns/iter (+/- 121)
test clone_from_small            ... bench:          22 ns/iter (+/- 0)
test clone_large                 ... bench:       2,083 ns/iter (+/- 56)
test clone_small                 ... bench:          33 ns/iter (+/- 0)
test insert_ahash_highbits       ... bench:       6,415 ns/iter (+/- 11)
test insert_ahash_random         ... bench:       7,067 ns/iter (+/- 18)
test insert_ahash_serial         ... bench:       6,472 ns/iter (+/- 36)
test insert_erase_ahash_highbits ... bench:      16,796 ns/iter (+/- 30)
test insert_erase_ahash_random   ... bench:      19,554 ns/iter (+/- 243)
test insert_erase_ahash_serial   ... bench:      16,612 ns/iter (+/- 34)
test insert_erase_std_highbits   ... bench:      42,649 ns/iter (+/- 101)
test insert_erase_std_random     ... bench:      44,372 ns/iter (+/- 10,354)
test insert_erase_std_serial     ... bench:      37,117 ns/iter (+/- 275)
test insert_std_highbits         ... bench:      18,120 ns/iter (+/- 1,112)
test insert_std_random           ... bench:      18,487 ns/iter (+/- 735)
test insert_std_serial           ... bench:      15,781 ns/iter (+/- 368)
test iter_ahash_highbits         ... bench:       1,830 ns/iter (+/- 90)
test iter_ahash_random           ... bench:       1,966 ns/iter (+/- 28)
test iter_ahash_serial           ... bench:       1,768 ns/iter (+/- 8)
test iter_std_highbits           ... bench:       1,928 ns/iter (+/- 48)
test iter_std_random             ... bench:       1,961 ns/iter (+/- 113)
test iter_std_serial             ... bench:       1,856 ns/iter (+/- 10)
test lookup_ahash_highbits       ... bench:       3,918 ns/iter (+/- 84)
test lookup_ahash_random         ... bench:       3,974 ns/iter (+/- 8)
test lookup_ahash_serial         ... bench:       3,879 ns/iter (+/- 180)
test lookup_fail_ahash_highbits  ... bench:       2,714 ns/iter (+/- 10)
test lookup_fail_ahash_random    ... bench:       3,644 ns/iter (+/- 70)
test lookup_fail_ahash_serial    ... bench:       2,739 ns/iter (+/- 31)
test lookup_fail_std_highbits    ... bench:      15,765 ns/iter (+/- 711)
test lookup_fail_std_random      ... bench:      15,766 ns/iter (+/- 63)
test lookup_fail_std_serial      ... bench:      14,495 ns/iter (+/- 93)
test lookup_std_highbits         ... bench:      15,337 ns/iter (+/- 707)
test lookup_std_random           ... bench:      15,370 ns/iter (+/- 408)
test lookup_std_serial           ... bench:      14,087 ns/iter (+/- 1,303)
```


## After

```
test clone_from_large            ... bench:       2,170 ns/iter (+/- 11)
test clone_from_small            ... bench:          22 ns/iter (+/- 0)
test clone_large                 ... bench:       2,137 ns/iter (+/- 60)
test clone_small                 ... bench:          34 ns/iter (+/- 0)
test insert_ahash_highbits       ... bench:       6,179 ns/iter (+/- 6)
test insert_ahash_random         ... bench:       6,888 ns/iter (+/- 53)
test insert_ahash_serial         ... bench:       6,318 ns/iter (+/- 18)
test insert_erase_ahash_highbits ... bench:      16,504 ns/iter (+/- 29)
test insert_erase_ahash_random   ... bench:      19,412 ns/iter (+/- 33)
test insert_erase_ahash_serial   ... bench:      15,993 ns/iter (+/- 41)
test insert_erase_std_highbits   ... bench:      44,060 ns/iter (+/- 108)
test insert_erase_std_random     ... bench:      44,821 ns/iter (+/- 1,881)
test insert_erase_std_serial     ... bench:      36,760 ns/iter (+/- 34)
test insert_std_highbits         ... bench:      18,071 ns/iter (+/- 933)
test insert_std_random           ... bench:      18,021 ns/iter (+/- 300)
test insert_std_serial           ... bench:      16,315 ns/iter (+/- 101)
test iter_ahash_highbits         ... bench:       1,854 ns/iter (+/- 37)
test iter_ahash_random           ... bench:       1,911 ns/iter (+/- 2)
test iter_ahash_serial           ... bench:       1,795 ns/iter (+/- 72)
test iter_std_highbits           ... bench:       1,901 ns/iter (+/- 33)
test iter_std_random             ... bench:       1,927 ns/iter (+/- 15)
test iter_std_serial             ... bench:       1,842 ns/iter (+/- 64)
test lookup_ahash_highbits       ... bench:       3,320 ns/iter (+/- 4)
test lookup_ahash_random         ... bench:       3,513 ns/iter (+/- 4)
test lookup_ahash_serial         ... bench:       3,113 ns/iter (+/- 4)
test lookup_fail_ahash_highbits  ... bench:       2,669 ns/iter (+/- 13)
test lookup_fail_ahash_random    ... bench:       3,633 ns/iter (+/- 181)
test lookup_fail_ahash_serial    ... bench:       2,653 ns/iter (+/- 24)
test lookup_fail_std_highbits    ... bench:      15,425 ns/iter (+/- 53)
test lookup_fail_std_random      ... bench:      15,293 ns/iter (+/- 85)
test lookup_fail_std_serial      ... bench:      13,399 ns/iter (+/- 83)
test lookup_std_highbits         ... bench:      15,322 ns/iter (+/- 303)
test lookup_std_random           ... bench:      15,372 ns/iter (+/- 97)
test lookup_std_serial           ... bench:      13,525 ns/iter (+/- 557)
```